lxml: try to remove some quirky PYPATH stuff
In RepositoryExternal.mk, a system python automatically gets $PYPATH
added to $PYTHONPATH, but the internal one does not - this doesn't make
sense.

Try to remove it for system-case by fixing the one case that relies on
it and for which it was introduced in commit
2ea0cb40b5aa3afc97788544b9010806321b8da1 to directly extend $PYTHONPATH
instead, which ought to work as long as it's not evaluated at global
scope and thereby avoids affecting all python invocations.

Change-Id: I4534f2be92b850dc01193cb1bb2e0a299a5152be
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/91748
Tested-by: Jenkins
Reviewed-by: Michael Stahl <michael.stahl@cib.de>