tdf#131490 sw btlr: fix handling of vertical text
Latin text goes from l->r, then t->b. If we rotate to the right, then we
get t->b, then r->l. Vertical text in vcl's Font means the individual
glyphs are painted in a way that looks "non-rotated" in the tbrl case.

btlr is not symmetric to this: if you rotate to the left, then Latin and
vertical text is handled the same way, i.e. there is no compensation at
a glyph level.

This means that as far as vcl is concerned, the Font's vertical flag has
to be true in the tbrl case, but no in the btlr one. Fix
SwFont::SetVertical() to do this, which means that rotating at a
character level or using the btlr text direction will result in the same
rendering for a one-liner text.

Regression from commit 2e013707144ffe0e2d45504f4a6cd2264e2d3f91 (sw btlr
writing mode: implement DOC filter, 2019-03-28).

Change-Id: I2619c77a3b2597dbf9feab6c7042e8d8c7454197
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/91780
Tested-by: Jenkins
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>