Revert "UITest: Actually time-out the wait"
This reverts commit eb702a8e845ad4ae86365f7f7744a588670bdfea.

Reason for revert:  With the above change, a hung soffice process would cause
the UITest that spawned it to fail, but that would not terminate the hung
soffice process.  This is especially problematic for Gerrit Jenkins builds,
where a leftover hung soffice process from a previous build will typically cause
unrelated subsequent builds to fail.  So:  For a manual developer build, it is
arguably better to leave the hung UITest in the hung state (both the python and
the soffice process) to be able to debug it.  And for an automated Gerrit
Jenkins build it is even vital that the hung UITest hangs the build, so that
Jenkins will eventually abort it and lode's bin/kill-wrapper will reliably kill
any remaining processes, incl. the hung soffice process.  (But of course it is
unfortunate that we now get less feedback from a hung UITest.)

Change-Id: I6cc577dda362d3cb71846decae9c2bb6bed90b6b
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/91717
Tested-by: Jenkins
Reviewed-by: Stephan Bergmann <sbergman@redhat.com>