tdf#131969: Fix reading SHORT Orientation value
<http://www.cipa.jp/std/documents/e/DC-008-2012_E.pdf> "CIPA DC- 008-Translation-
2012: Exchangeable image file format for digital still cameras: Exif Version 2.3"
documents that the Orientation tag 0x0112 expects a count of 1 values of type SHORT
(16 bit), and details that values <= 4 bytes are stored in the Value Offset field
always using bytes starting from the left of the field.

This is a regression introduced with 20ce9f46046e716617cf62794bf76eb132397839 "Avoid
-fsanitize=misaligned-pointer-use".  That commit had wondered why the original code
had used OSL_SWAPWORD instead of OSL_SWAPDWORD when reading and writing such
orientation values.  It turns out that that original code had happened to work
correctly when processing either little or big endian data on a little endian
machine.  (Though it would have worked incorrectly when processing either little or
big endian data on a big endian machine.)  And with
20ce9f46046e716617cf62794bf76eb132397839, the code worked when processing little
endian data on a little endian machine, but failed when processing big endian data
on a little endian machine, as is the case for tdf#131669 on e.g. x86_64.

(read32 has become unused and is thus removed.)

Change-Id: I7992629048ac44c00ee703c75164f3d094773244
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/91881
Tested-by: Jenkins
Reviewed-by: Stephan Bergmann <sbergman@redhat.com>
Reviewed-by: Tomaž Vajngerl <quikee@gmail.com>