sw: crashtesting: fix assert on ooo95711-1.odt
Document is special and contains deletion that contains only a table, no
paragraphs before or after it.  The assertion is wrong in this case.
It doesn't matter if the atParaEndNode is incremented because it's going
to be on an end node either way.

(regression from 51c97e5e3ae88c43e611a4e6a61e375c66d5d007)

Change-Id: I3fe82f0ed0b7b213df036bb8026dc209c00123de
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/91873
Tested-by: Jenkins
Reviewed-by: Michael Stahl <michael.stahl@cib.de>