sw: fix fly at-char deletion API behaviour change
245b52ee39d03c62998acba87405ead7afc51d71 changed at-char anchored fly
selection.

WollMux calls setString("") to remove one character in the first
paragraph and a fly anchored at the start of the body text is deleted
due to the new IsAtStartOfSection() check.

It would be possible to treat deletion via API differently than via UI,
as there is already a flag ExcludeFlyAtStartEnd but it would require
passing the flag through 10 functions and also to store it in
SwUndoDelete...

The main intent of the IsAtStartOfSection() check was that Ctrl+A should
delete every fly; this can be achieved by checking that everything was
selected, so that selections only inside the first/last paragraph don't
delete the flys at the edges.

Change-Id: Id62e7d99a10b42eaecea5564372d29b6c43e3f7e
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/91993
Tested-by: Jenkins
Reviewed-by: Michael Stahl <michael.stahl@cib.de>