NFC writerfilter: remove pointless asserts
In an age where GDB (or whatever other platform debugger
you prefer) is easy to use, there is no point in
asserting something that is going to crash anyway, is there?

Asserting is only good in these cases if you follow it using
an _if_ statement. Noel informed me that it can also be used
to silence false positive coverity warnings.

Change-Id: I5a5cb7a22019768ec2807f6918d4a8ebb51194de
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/92049
Tested-by: Jenkins
Reviewed-by: Noel Grandin <noel.grandin@collabora.co.uk>