fix "NFC writerfilter numbering: remove redundant code"
Earlier, in LO 7.0, I made an error in
commit 13d8a32ffb6fa5593b75f1c9453fb261d8882ce8, where
I failed to notice the difference between setLineId and setNumId,
and removed setNumId as what I thought was duplicate code.
(Thanks for pointing that out Mike Kaganski!)

Reviewing the relevant code, setNumId seems to effectively be
redundant after all. Removing it and just using the LineId value.

There is a slight difference in the code flow, since numId was
only set if the list didn't exist, but that appears to
be irrelevant. Lists are not parsed/created until after
styles are finished, so !pList.get() is implied by IsStyleSheetImport().

Change-Id: I2575966c8f1781bb278c787a2928d2b459867940
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/92125
Tested-by: Jenkins
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>