PDF export: fix handling of page sizes larger than 508 cm
The value of these coordinates are not allowed to be larger than 14 400,
and Adobe Reader complains about them.

Use UserUnit to declare in case we won't work with points anymore, but
with a larger unit. This will mean UserUnit=2 in practice, since e.g.
Draw has is page size limited to 600x600cm, so larger values won't
happen, at least not for now.

Change-Id: I8ee159f2571f4070aded85388792a215de86f7ff
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/92383
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>
Tested-by: Jenkins