tdf#132642 sw layout: try2 emulate table kept-with-next not splitting
This adjusts my LO 5.2 code for tdf#91083 that tried to emulate the
table's keep-with-next property which doesn't have a matching
counterpart in MS formats.

I always confused myself trying to understand what my year-long coding
attempt was trying to do. This seems much understandable, and efficient.
The big clue was that it affected non-MS formats - which was unintended.

Change-Id: I7886e52430cb34799e25f7fcf73500e28bbe2a55
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/93443
Tested-by: Jenkins
Reviewed-by: Justin Luth <justin_luth@sil.org>
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>