Resolves: tdf#132581 allow disabling the auto-group RadioButton feature
the older way of doing things is to dig around contiguous radiobuttons to see
what is part of the group, while the contemporary one is to explicitly use
group() to set the members. A problem of the legacy way of doing it is that the
results during setup and teardown of a collection of radiobuttons can be
misleading.

Change-Id: I6c590d0fb6fc6db3f5e5ba4397a62930a5c8f3f0
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/93567
Tested-by: Jenkins
Reviewed-by: Caolán McNamara <caolanm@redhat.com>