tdf#132267 vcl: fix missing scrollers with non-native rendering
Regression from 5a4b49d3bccfb6d4e4451f8cf9a794f9df59874e (tdf#125415 vcl
menu floating window: avoid flicker, 2019-05-21) the problem was that
the clip region was set on the buffer, not on the render context. This
means the original clip was used to determine what gets copied from the
buffer to the screen, so the scroller arrows were not rendered.

Change-Id: Id173e6333721891798da58baf2092f4cd21a62ba
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/93642
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>
Tested-by: Jenkins