Related: tdf#129916 svx: improve shadow size of custom shapes
There are multiple problems with this bug document, the first is that
the shadow primitive got the default position (0) of the owning shape.
Do it the same way as commit ccc95106faa47ccdc4f69bfe468f66624cb8c61c
(tdf#130058 Import shadow size., 2020-04-14) did for graphic objects.

This requires constructing a transform matrix in
ViewContactOfSdrObjCustomShape::createViewIndependentPrimitive2DSequence(),
include position and size in that as a start.

Change-Id: Ia51df555c1984971afe7c52ba3f2658099a4e7b3
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/93771
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>
Tested-by: Jenkins