tdf#129916 svx: fix origin of scaled shadow
We assumed that the top left corner is the origin for scaling, but that
is not necessarily the case. The intention is that the shadow direction
can be determined from its x and y offset, and the origin is the
opposite corner of the shape.

Change-Id: I6759302767d20739b6e2be79d379740dd06f70f5
(cherry picked from commit a1dde1a85a8d1c54521a0ac6310571ffcdd4bb5a)
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/93974
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>
Tested-by: Jenkins