tdf#133000 writerfilter: apply non-contradictory num-style
This fixes 7.0 regression exposed by tdf#131321's
commit e33495fc5f4adb709422bd1719ec46a382728733

This patch version heavily depends on recent changes,
but could easily be adapted to older ways.
(See previous versions in gerrit.)

Because LO can now find the numbering rule through
the para-style definition, it doesn't bother creating it
as a direct paragraph property anymore.
So that "directness" has to be specified during the import.

This problem manifested itself in two ways.
1.) In SW, the direct numbering was lost and it simply
    inherits from the para-style. If numbering is now turned off
    in the para-style, then the paragraph loses it too.
2.) Because the numbering was considered to be
    defined by the paragraph style, the para-style-indents
    had higher priority than the numbering-indents.
    When numbering is defined directly on the paragraph,
    it has higher priority than the para-style-indents.

Change-Id: I04c3944de8a91d9f253791fbd05d6324a8b7a9da
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/94365
Tested-by: Jenkins
Reviewed-by: Justin Luth <justin_luth@sil.org>