package: ODF export: fix position of manifest:start-key-generation child
META-INF/manifest.xml[8,137]:  Error: element "manifest:start-key-generation" was found where no element may occur

The problem is that in the schema, manifest:start-key-generation
precedes key-derivation but LO writes it in swapped order.

This is apparently a workaround for a bug in OOo < 3.2 ODF import and
specific to ODF >= 1.2 because the element didn't exist in ODF 1.1; as
it turns out we don't care about OOo < 3.4 being able to load encrypted
ODF 1.2 documents anway since commit
8075e3fb61f3e29d9aae39b5fffceb3ca9c34748 "tdf#114939 officecfg,sfx2:
always use AES/SHA256 in ODF 1.2".

Change-Id: I3b62e3eba719f13e39fb24bf9792f764b24ccaeb
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/94413
Tested-by: Jenkins
Reviewed-by: Michael Stahl <michael.stahl@cib.de>