sw HTML import: ignore hidden text frames
Use-case: HTML-copy 2 lines from a browser from a mattermost chat, paste
it into Writer. The content is not visible at all, because there are 5
invisible full-page textframes covering the actual content.

The browser has no problem with showing the content as these divs are
all invisible. Writer does not support hiding textframes, so just ignore
them on import for now.

Change-Id: I36a807d46f5f8348239e4693ec85da49a2205854
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/94500
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>
Tested-by: Jenkins